requirements.txt
pandas
python-dotenv
torch
transformers
accelerate
langchain
langchain-community
